requirements.txt
streamlit
pypdf
faiss-cpu
sentence-transformers
openai
python-dotenv
